New index() method to find handlers
The handlers list contains tuples of condition and actual handler.

It is very useful to manipulate it to add new types.

However when it is desirable to replace an handler, it might be difficult to find it, without
directly going into the code, and the ordering of the handlers might change.

This API allows the users to find handlers, and perhaps replace them, or know their position
to add other handlers before or after that position.