Run all tests under python 3.7 with future annotations
Under python 3.7, re-run all the tests importing annotations from future, so I have an idea of
what fails.

It just copies the same test files adding the import, and modifies the entry point for the test
to load them for python 3.7 and greater. It will have to be modified when it becomes the default.