Make the list loader return a TypedloadException when loading a non-iterable
It was raising a TypeError.

It took me so long to do because the except branch was also being executed
in case the TypeError was coming from a lower function, and then it was getting
thrown away to make a new exception, but since it was already also a TypedloadException
type, that was incorrect and was losing information, so the tests were correctly failing.

This only makes a new exception object if the raised one is not already of TypedloadException.

Ah, the joys of multiple inheritance ☺