Do not load dict into set/dict/tuple
This solves an issue when loading something into

`Union[List[str], Dict[str, str]]`

or loading into a NamedTuple or similar thing.

Because a dict is iterable it could be loaded as a list and
the result would be a list of the keys in the dictionary.

This is not what I want so by making the load fail
the union loader will try other options.

Because Unions are not ordered the code before this commit
will return different results randomly.