failonextra triggers a failure if fields are dropped in mangling
In name mangling now fields which use the python name instead of the mangled name get dropped and not passed forward.

This means that what used to work, passing the python unmangled name in the dictionary, no longer works.

However, the field was dropped silently. This patch makes the drop raise an exception if failonextra is True.