Allow duck typing for attr loading
attr handler was checking that the value is a dictionary.

While this is the intention, removing the explicit check
allows for dictionary-like classes to be used instead.

This is good, I don't want to be like those libraries
that check that paths must be strings and refuse pathlib.Path :D