Union raises immediately when a non typedload exception is found
Basically all handlers are required to only raise subclasses of TypedloadException
at all times and in any case.

A recently met bug was hidden and hard to find because within a union the original
exception was hidden among the other, and then everything was crashing when the
non-typedload exception was inspected to provide information on the issue.

This prevents the later stage failure and fails right away instead.