contributing file gives freedom to upgrade license
After an email exchange with rms (Richard Stallman), who explained to me why it
is important to use GPL3+ license, to allow compatibility, and my worries about
the future of the FSF, I decided to do this compromise:

the version for the license is 3, but should other versions be published I
promise to all contributors that I will relicense as 3 and 4 (or 3, 4 and 5) if
the new licenses do not alter the current spirit of the license.

I change the contributing file so the contributors have to accept this rule and
are aware that their contribution might one day become gpl3+ if such license
should be published.