Use get_type_hints if pep563 is enabled
It's slow, doesn't support nesting, but people seem to use it.

Probably they use it because they never actually use the type hints so it
doesn't matter to them.

This adds a flag to load types using `get_type_hints` which in turn calls
`eval`.

Because this will be superseeded by https://peps.python.org/pep-0649/ the flag
is starting its life as "deprecated", so I will be free to remove it whenever I
want.