Make list loader faster
Instead of calling load() call the handler directly and skip a
function call.

We lose the annotation information because iterators from comprehensions
can't be extracted, and a for loop is much slower than a comprehension.

This will be solved when dropping everything prior to python3.8 when
it will be possible to extract the point when the error occurred in
the iterator.

Until then, basically the list is loaded again but calling load() and
with the annotations if an exception comes, so that the exceptions still
have the details needed to track the problem.

The idea is that exceptions are much rarer than normal condition and so
it makes sense to slow down the exception handling to make the normal
loading much faster.